refactor : 폴더 카테고리 렌더링 컴포넌트 FolderMain 으로 이동
